Moved NewModuleNameModalProps and NewModuleNameModal to src/modal/NewModuleNameModal.tsx.
